v4l: max9286: Use internal automatic frame synchronization
The MAX9286 can send frame synchronization packets to synchronize
multiple sources. Frame synchronization can be external (driven by the
SoC) or internal (generated by the MAX9286).

The MAX9286 frame sync (FSYNC) input is connected to the SoC, but in the
current setup the SoC doesn't generate any signal on that pin. Configure
internal automatic frame sync unconditionally to reflect that.

Signed-off-by: Laurent Pinchart <laurent.pinchart+renesas@ideasonboard.com>