v4l: max9286: update how streams are started
It dose not work to start cameras in sequence, the frame synchronisation
lock is never set. If this is due to other configurations or due to
hardware limitations are unclear. But starting just the camera
associated with the requested stream at least allows for capturing when
more then one camera is attached.

Note: This will allow capturing from any attached camera, not just at
the same time.

Signed-off-by: Niklas Söderlund <niklas.soderlund+renesas@ragnatech.se>