media: i2c: max9286: Configure local ack on I2C during camera probe
The MAX9286 initialises with auto-acknowledge enabled by default. This
means that if multiple MAX9286 devices are connected to an I2C bus, an
uninitialised MAX9286 might respond with ack's for devices such as
MAX9271 which are being probed before the second MAX9286 is fully
initialised.

To prevent this issue, ensure that the I2C auto acknowledge is disabled
using the default configuration of the I2C configuration register, with
the auto-ack disabled early on.

However, RDACM20 cameras without an MCU or with faulty MCU's will need
auto-ack enabled on the MAX9286 before the forward channel is enabled on
the MAX9271 serializer.

Ensure that the Local Acknowledge is enabled at the MAX9286 side of the
GMSL link during camera probe sequences. This allows us to successfully
configure cameras which do not have their forward configuration channel
enabled, but are actively listening to I2C configuration writes.

I knowingly left this code such that we 'disabled' local ack in two
locations. This allows us to configure a different speed later if needed
whilst still retaining the correct local ack disable sequence:

 * max9286_probe() : Configure default I2C parameters, disable localack
 * max9286_setup() : Enable local ack. We can configure DT speeds here if
                     desired. Speed configuration must then be stored in
                     the max9286_device structure in that instance.
 * max9286_init()  : After the cameras have probed, re-configure I2C
                     with local ack disabled.

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>

---
v2:
 - Disable local ack on error path too