media: i2c: rdacm20: Configure I2C and GMSL link before identifying device
Ensure that we make a full attempt at configuring both the GMSL link and
the I2C link before we try to read from the device.

This is vital in ensuring link stability and correct configuration to be
able to read the device ID correctly, and program the new device
address.

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>