rcar-vin: sort compatibility strings in bindings documentation
The list of compatibility strings where not properly sorted, fix this.

The compatibility string "renesas,rcar-gen3-vin" was added before the
Gen3 driver code was added and is not possible to use as each SoC in the
Gen3 series require SoC specific knowledge to function. Remove it before
it is added to any device tree descriptions.

Signed-off-by: Niklas Söderlund <niklas.soderlund+renesas@ragnatech.se>