[PoC] arm64: dts: renesas: r8a77995-draak: Add ADV7511
Provide a device node for the ADV7511 as found on the Draak D3 platform.

The ADV7511 is connected to the DU through a parallel mux chip, and is
configurable in hardware whether it is connected to LVDS0 or LVDS1.

This is a development process only patch, to enable the ADV7511 on the
I2C bus, and allow verification of the I2C address configuration.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>