drm: adv7511: Register the ADV7511 fixed address mapping
The ADV7511 programming guide specifies that a fixed I2C address must be
programmed into register 0xF9 of the main map which does not conflict
with other devices on the bus.

Register an I2C client to correspond to the default address used by the
hardware to ensure that it is non-conflicting, and also allow it to be
changed programatically through the device tree node.

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>