rcar-vin: sync which hardware buffer to start capture from
When running consecutive capture sessions the hardware don't always
start delivering frames captured from the VnMB1 buffer. This is
problematic if we need to capture to a vb2 buffer two times at different
offsets to support SEQ_TB/BT capering. Solve this by dropping up to max
2 fields at capture start to always start capture from VnMB1.

As an added benefit with this change buffers are always returned to
userspace in the same order they are queued.

Signed-off-by: Niklas Söderlund <niklas.soderlund+renesas@ragnatech.se>