dt-bindings: media: adv7604: Fix slave map documentation
The reg-names property in the documentation is missing an '='. Add it.

Fixes: 9feb786876c7 ("media: dt-bindings: media: adv7604: Extend
bindings to allow specifying slave map addresses")

Signed-off-by: Kieran Bingham <kieran.bingham@ideasonboard.com>

---
v2:
 - Commit title changed to prefix as "dt-bindings: media:"

If this is collected through a DT tree, I assume therefore this will be
fine, but if it is to go through the media-tree, please update as
necessaary to prevent the redundant dual "media:" tagging.

(I'll leave it to the maintainers to decide whose tree thise should go
through)

Thanks

Kieran