drm: rcar-du: Move routing configuration
The rcar_du_crtc_setup() currently configures the group routing. This
can now be handled directly by the group. Introduce a new flag
'active_changed' to track if the active state of any CRTC in the group
changes it's active flag, and in the event of a change update the
routing.

The routing will only be updated if there is at least one active CRTC.

This commit changes the sequence of register writes to the DU. The
routing is now configured *before* rcar_du_crtc_setup() is called, and
as such is set before setting the background to black and disabling the
display.

Signed-off-by: Kieran Bingham <kieran.bingham+renesas@ideasonboard.com>